moving some stuff into utils folder that will be needed between modules
